Adding Initially Pre-created Files
I am adding the files that I have been using within folders that were not within a repository. Since they are what this repository is for, they will be added here. This includes the original folder for the "OBS-JUXT.zip" file. Next I will be adding the ascii-doc that I am finishing writing.